remove deprecated components
