+++ client/src/pages/home.jsx
@@ -1,19 +1,13 @@
-import React from 'react';
-import '../stylings/home.css';
-import Footer from '../components/footer';
-import Search from '../components/search';
-import Courses from '../components/courses';
+import React from "react";
+import "../stylings/home.css";
+import Footer from "../components/footer";
+import Search from "../components/search";
 
 const Home = () => {
   return (
     <div id="home">
       <Search />
-      <div>
-        <Courses />
-      </div>
-      <div>
-        <Footer />
-      </div>
+      <Footer />
     </div>
   );
 };
